+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
 detailed information.
 
+## [1.53.0]
+
+### Added
+
+- Add the `defaultNodejsVersion` attribute to UNIX user.
+- Add the `cpuLimit`, `appRoot` and `isNamespaced` attributes to Passenger app.
+
 ## [1.52.0]
 
 ### Added